Notes CRUD
Add CRUD endpoint for the Notes.

The validation is a bit janky since the endpoints are updating single properties and fluentValidation works with objects. However I wanted to keep duplication to a minimum.

Also opted for a base repository class to help streamline working with the mongo driver

I unfortuantely commited the appSettings.development and launch settings so will need to untrack them