Implement pagination
Quite pleased that I learned how to do cursor pagination.

The benefit here is that adding/deleting on the frontend should cause any nasty pagination bugs.

Unlike the traditional skip and limit which works better for specific page selection but given my app is going for a load more approach, this cursor approach is definitely better